Add zlib-ng crc32 calculation in addition to the legacy pzs-ng's crc32.
The autotools setup is updated, and we add zlib-ng setup to it.
Allow also custom paths by specifying --with-lzlib_ng_path and --with-lzlib_ng_include using absolute paths.
Now configure autodetects pclmulqdq and only when support tries to check for zlib-ng.
Having the macro defined, we call zlib-ng instead of the legacy pzs-ng function.
Instructions are added in INSTALL.